Extracted Toggler in shared folder
