+++ src/h5web/visualizations/shared/Toggler.module.css
@@ -1,0 +1,16 @@
+.toggler {
+  padding: 0 0.75rem;
+  composes: btn-clean from global;
+  display: flex;
+  align-items: center;
+  justify-content: left;
+}
+
+.toggler:hover,
+.toggler[aria-checked='true'] {
+  background-color: var(--primary-dark-bg);
+}
+
+.togglerLabel {
+  margin-left: 0.5rem;
+}
